feat: refactor job components and introduce new job details and status badge components
- Updated import paths for JobDetails component.
- Added new JobDetails, JobStatusBadge, and job-columns components for better job management.
- Refactored JobsTable to utilize new job columns.
- Removed obsolete job layout and loading components.
- Cleaned up unused job input and detail section components.
