+++ web-app/src/app/app/agents/[agentId]/jobs/@right/components/job-details.tsx
@@ -2,7 +2,7 @@ import { ScrollArea } from "@radix-ui/react-scroll-area";
 import { useFormatter, useTranslations } from "next-intl";
 import Markdown from "react-markdown";
 
-import JobStatusBadge from "@/app/jobs/components/job-status-badge";
+import JobStatusBadge from "@/app/agents/[agentId]/jobs/@right/components/job-status-badge";
 import { JobWithRelations } from "@/lib/db/services/job.service";
 import { cn } from "@/lib/utils";
