+++ web-app/src/app/app/agents/[agentId]/jobs/@right/components/job-status-badge.tsx
